Finish implementing TofinoPortManager class (#26)
- Extract tofino_port_manager.cc from tofino_sde_target.cc.

- Declare methods in tofino_port_manager.h to be virtual, so they can be overridden by the TofinoPortManagerMock.

- Implement TofinoPortManagerMock class.

- Update TofinoChassisManagerTest to work with TofinoPortManager.

- Update Bazel build files.
